# api/src/main/resources/openapi/arduino-thermo-hygrometer-api.yaml
openapi: 3.1.0
info:
  title: Arduino Thermo Hygrometer API.
  description: A Java Web API for an Arduino Thermo Hygrometer IoT device.
  contact:
    name: Burhan Mohammad Sarfraz
    email: burhan.mohammad.sarfraz@outlook.com
  license:
    name: MIT license
    url: https://mit-license.org/
  version: 1.0.0-SNAPSHOT

servers:
  - url: http://localhost:5000
    description: Local development server.

tags:
  - name: Battery.
    description: Battery status of arduino device.
  - name: Temperature.
    description: Temperature measured by arduino device.
  - name: Humidity.
    description: Humidity measured by arduino device.

paths:
  /api/temperatures:
    post:
      tags:
        - Temperature.
      summary: Create measured temperature.
      operationId: createTemperature
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureDto"
        required: true
      responses:
        "201":
          description: Temperature created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureDto"
        "400":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/BadRequest"

  /api/humidities:
    post:
      tags:
        - Humidity.
      summary: Create measured humidity.
      operationId: createHumidity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumidityDto"
        required: true
      responses:
        "201":
          description: Humidity created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumidityDto"
        "400":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/BadRequest"

  /api/batteries:
    post:
      tags:
        - Battery.
      summary: Create remaining battery.
      operationId: createBattery
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatteryDto"
        required: true
      responses:
        "201":
          description: Battery created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatteryDto"
        "400":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/BadRequest"

  /api/temperatures/{id}:
    get:
      tags:
        - Temperature.
      summary: Retrieve measured temperature by id.
      operationId: getTemperatureById
      parameters:
        - in: path
          name: id
          description: Identifier for measured temperature.
          required: true
          schema:
            type: string
            format: uuid
            example: "86644f52-f3d4-4285-9f2f-4d2ee4c15035"
      responses:
        "200":
          description: Temperature found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"
    delete:
      tags:
        - Temperature.
      summary: Delete measured temperature by id.
      operationId: deleteTemperatureDtoById
      parameters:
        - in: path
          name: id
          description: Identifier for measured temperature.
          required: true
          schema:
            type: string
            format: uuid
            example: "86644f52-f3d4-4285-9f2f-4d2ee4c15035"
      responses:
        "204":
          description: Temperature deleted.
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/temperatures/timestamp:
    get:
      tags:
        - Temperature.
      summary: Retrieve measured temperature by timestamp.
      operationId: getTemperatureByTimestamp
      parameters:
        - in: query
          name: timestamp
          description: Timestamp for measured temperature.
          required: true
          schema:
            type: string
            format: date-time
            example: "2017-07-21T17:32:28Z"
      responses:
        "200":
          description: Temperature found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"
    delete:
      tags:
        - Temperature.
      summary: Delete measured temperature by timestamp.
      operationId: deleteTemperatureByTimestamp
      parameters:
        - in: query
          name: timestamp
          description: Timestamp for measured temperature.
          required: true
          schema:
            type: string
            format: date-time
            example: "2017-07-21T17:32:28Z"
      responses:
        "204":
          description: Temperature deleted.
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/temperatures/date:
    get:
      tags:
        - Temperature.
      summary: Retrieve measured temperatures by date.
      operationId: getTemperaturesByDate
      parameters:
        - in: query
          name: date
          description: Date for measured temperatures.
          required: true
          schema:
            type: string
            format: date
            example: "2017-07-21"
      responses:
        "200":
          description: Temperatures found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemperatureDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/humidities/{id}:
    get:
      tags:
        - Humidity.
      summary: Retrieve measured humidity by id.
      operationId: getHumidityById
      parameters:
        - in: path
          name: id
          description: Identifier for measured humidity.
          required: true
          schema:
            type: string
            format: uuid
            example: "86644f52-f3d4-4285-9f2f-4d2ee4c15035"
      responses:
        "200":
          description: Humidity found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumidityDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"
    delete:
      tags:
        - Humidity.
      summary: Delete measured humidity by id.
      operationId: deleteHumidityDtoById
      parameters:
        - in: path
          name: id
          description: Identifier for measured humidity.
          required: true
          schema:
            type: string
            format: uuid
            example: "86644f52-f3d4-4285-9f2f-4d2ee4c15035"
      responses:
        "204":
          description: Humidity deleted.
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/humidities/timestamp:
    get:
      tags:
        - Humidity.
      summary: Retrieve measured humidity by timestamp.
      operationId: getHumidityByTimestamp
      parameters:
        - in: query
          name: timestamp
          description: Timestamp for measured humidity.
          required: true
          schema:
            type: string
            format: date-time
            example: "2017-07-21T17:32:28Z"
      responses:
        "200":
          description: Humidity found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumidityDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"
    delete:
      tags:
        - Humidity.
      summary: Delete measured humidity by timestamp.
      operationId: deleteHumidityByTimestamp
      parameters:
        - in: query
          name: timestamp
          description: Timestamp for measured humidity.
          required: true
          schema:
            type: string
            format: date-time
            example: "2017-07-21T17:32:28Z"
      responses:
        "204":
          description: Humidity deleted.
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/humidities/date:
    get:
      tags:
        - Humidity.
      summary: Retrieve measured humidities by date.
      operationId: getHumiditiesByDate
      parameters:
        - in: query
          name: date
          description: Date for measured humidities.
          required: true
          schema:
            type: string
            format: date
            example: "2017-07-21"
      responses:
        "200":
          description: Humidities found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumidityDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/batteries/{id}:
    get:
      tags:
        - Battery.
      summary: Retrieve a battery by id.
      operationId: getBatteryById
      parameters:
        - in: path
          name: id
          description: Identifier for battery.
          required: true
          schema:
            type: string
            format: uuid
            example: "86644f52-f3d4-4285-9f2f-4d2ee4c15035"
      responses:
        "200":
          description: Battery found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatteryDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"
    delete:
      tags:
        - Battery.
      summary: Delete battery by id.
      operationId: deleteBatteryDtoById
      parameters:
        - in: path
          name: id
          description: Identifier for battery.
          required: true
          schema:
            type: string
            format: uuid
            example: "86644f52-f3d4-4285-9f2f-4d2ee4c15035"
      responses:
        "204":
          description: Battery deleted.
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/batteries/timestamp:
    get:
      tags:
        - Battery.
      summary: Retrieve a battery by timestamp.
      operationId: getBatteryByTimestamp
      parameters:
        - in: query
          name: timestamp
          description: Timestamp for battery.
          required: true
          schema:
            type: string
            format: date-time
            example: "2017-07-21T17:32:28Z"
      responses:
        "200":
          description: Battery found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatteryDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"
    delete:
      tags:
        - Battery.
      summary: Delete battery by timestamp.
      operationId: deleteBatteryByTimestamp
      parameters:
        - in: query
          name: timestamp
          description: Timestamp for battery.
          required: true
          schema:
            type: string
            format: date-time
            example: "2017-07-21T17:32:28Z"
      responses:
        "204":
          description: Battery deleted.
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

  /api/batteries/date:
    get:
      tags:
        - Battery.
      summary: Retrieve batteries by date.
      operationId: getBatteriesByDate
      parameters:
        - in: query
          name: date
          description: Date for batteries.
          required: true
          schema:
            type: string
            format: date
            example: "2017-07-21"
      responses:
        "200":
          description: Batteries found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BatteryDto"
        "404":
          $ref: "https://api.swaggerhub.com/domains/smartbear-public/ProblemDetails/1.0.0#/components/responses/NotFound"

components:
  schemas:
    TemperatureDto:
      type: object
      properties:
        registeredAt:
          type: string
          format: date-time
        temp:
          type: number
          format: double
          minimum: -55.00
          maximum: 125.00
      example:
        registeredAt: "2017-07-21T17:32:28Z"
        temp: 100.00
      required:
        - registeredAt
        - temp

    HumidityDto:
      type: object
      properties:
        registeredAt:
          type: string
          format: date-time
        airHumidity:
          type: number
          format: double
          minimum: 20.00
          maximum: 90.00
      example:
        registeredAt: "2017-07-21T17:32:28Z"
        airHumidity: 70.00
      required:
        - registeredAt
        - airHumidity

    BatteryDto:
      type: object
      properties:
        registeredAt:
          type: string
          format: date-time
        batteryStatus:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
      example:
        registeredAt: "2017-07-21T17:32:28Z"
        batteryStatus: 50
      required:
        - registeredAt
        - batteryStatus